+++ test/unit/fuzz-tools.test.js
@@ -137,6 +137,7 @@ import * as endoV136 from '../../lib/endo-v136.js';
 import * as idV137 from '../../lib/id-v137.js';
 import * as obV138 from '../../lib/ob-v138.js';
 import * as gynV139 from '../../lib/gyn-v139.js';
+import * as pedsV140 from '../../lib/peds-v140.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -219,6 +220,7 @@ const MODULES = {
   'id-v137.js': idV137,
   'ob-v138.js': obV138,
   'gyn-v139.js': gynV139,
+  'peds-v140.js': pedsV140,
   'ems-v149.js': emsV149,
 };